requirements.txt
Pillow==7.1.1
Flask==1.1.2
